aawaz/requirements.txt
asgiref==3.5.2
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
defusedxml==0.7.1
diff-match-patch==20200713
dj-rest-auth==2.2.5
Django==3.2
django-allauth==0.51.0
django-cors-headers==3.13.0
django-import-export==3.0.1
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
firebase-admin==6.0.1
google-api-core==2.11.0
google-api-python-client==2.68.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
grpcio-status==1.51.1
httplib2==0.21.0
idna==3.4
jwt==1.3.1
MarkupPy==1.14
msgpack==1.0.4
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.3.0
proto-plus==1.22.1
protobuf==4.21.10
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tablib==3.2.1
uritemplate==4.1.1
urllib3==1.26.13
xlrd==2.0.1
xlwt==1.3.0
